requirements.txt
torch
numpy
torchvision
pillow
fastapi[standard]
python-dotenv
boto3
opencv-python
onnxruntime
sqlalchemy
passlib[bcrypt] 
python-jose
psycopg2-binary
pydantic_settings